Moved functions to new module for clarity
